+++ backend/.dockerignore
@@ -25,5 +25,3 @@ tests/
 
 # PCX
 pcr-backup
-frontend/
-tests/
