+++ backend/tests/alert/test_alert.py
@@ -10,11 +10,12 @@ from django.urls import reverse
 from django.utils.dateparse import parse_datetime
 from options.models import Option
 from rest_framework.test import APIClient
+from tests.courses.util import create_mock_data
 
 from alert import tasks
 from alert.models import SOURCE_PCA, Registration, RegStatus, register_for_course
 from courses.models import StatusUpdate
-from courses.util import create_mock_data, get_or_create_course_and_section
+from courses.util import get_or_create_course_and_section
 
 
 TEST_SEMESTER = "2019A"
